Move the dynamic validator out of api/
